Fix delete api

+++ api/v1/expenses/delete_expense.py
@@ -6,15 +6,19 @@ cursor = create_cursor(conn)
 
 def delete_expense(id_spesa):
     try:
-        # Esegui la query per eliminare la riga corrispondente all'id_spesa specificato
+        # Controlla se la spesa esiste prima di eliminarla
+        cursor.execute("SELECT id FROM spese WHERE id = %s", (id_spesa,))
+        expense = cursor.fetchone()
+
+        if not expense:
+            return jsonify({"error": "Expense not found"}), 404
+
+        # Se esiste, elimina la spesa
         cursor.execute("DELETE FROM spese WHERE id = %s", (id_spesa,))
-        
-        # Conferma la transazione e applica le modifiche al database
         conn.commit()
-        
+
         return jsonify({"message": "Deleted successfully!"}), 200
     except Exception as e:
         print("Error!:", str(e))
-        # In caso di errore, annulla le modifiche e restituisci un messaggio di errore
         conn.rollback()
         return jsonify({"error": "Impossible to remove the expense"}), 500
